feat, WIP: moved new code to dedicated file in EvoNet
